Add many builtins; make it easier to add methods to PythonLikeType
To make it simpler to add methods to PythonLikeType, addMethod now
has an `addMethod(String methodName, Method method)` overload, which
determines the function signature from the method return and parameter
types.

Because Python ints and floats can be equal to one-another (and thus
can refer to a common key in a dict), their hash function must return
the same value when they are equal. The exact function is described in
https://docs.python.org/3/library/stdtypes.html#hashing-of-numeric-types.

Fix a bug with detecting native objects. dis.code_info cannot
reliable be used to determine if an object is native
(i.e it throws an exception for objects even if they are
not native). Instead, the proper way to detect native objects
are to get their modules, and check if the module loader is
the ExtensionFileLoader.

Set "extra" attributes when converting a Python object to
its equalivant Java object (via its __dict__ attribute).

Builtins added:

- ascii
- bin
- callable
- chr
- delattr
- divmod
- enumerate
- format
- getattr
- globals
- hasattr
- hash
- hex
- id
- input
- isinstance
- issubclass
- map
- object
- oct
- ord
- pow
- range
- setattr