Support BUILD_SLICE; fix parameter typing for PythonLikeFunction
Previously, PythonLikeFunction was typeflowed like any other function.
This has an issue: PythonLikeFunction arguments are the list of
positional arguments + the map of named arguments. In other words,
the arguments to PythonLikeFunction are not the arguments of the
function, but grouped arguments of the function, thus their types
(List and Map) are not the parameter types of the function.

This caused an issue where if the function was compiled as a
PythonLikeFunction, the first argument is always read as a list,
and the second argument a dict. Because the typing was incorrect,
an ClassCastException might be raised (in particular, when it tries
to invoke a virtual method on list).

To fix it, when creating the initial stack metadata, PythonLikeFunction
goes into its own branch, where all local variables are assigned
base-object as their types (instead of only the first two
parameters getting list and dict as their types).

Newly supported opcodes:

- BUILD_SLICE