Fix IndexOutOfBound exception when a virtual method with no parameters
Despite being very bad pratice, it is valid to have a python function
in a class (that is not a static or class method) without any
parameters (in particular, self).

The old code getInitialStackMetadata code assume there was
always at least one parameter (for virtual methods), but that
is not the case. Now, it check the method parameter count
before setting the value source for the first parameter
(in particular, no value source is set if the method has
no parameters).