Fix cell and free variables in generators; sanitize variable names
Previously, the generator logic assumed that
co_cellvars and co_freevars was included in
co_varnames. This was false, thus when cell
variables or free variables are used in generators,
an exception is thrown since writing a local variable
overrides the cell variable (since both the local
variable and cell variable shared the same slot
due to the above logic).

The generator logic was corrected so cell variables
and free variables are placed after local variables.
To investigate: if this is also the case outside
of generators.

For functions corresponding to code objects
(often generated by comprehensions), they may
have parameter names of the form ".N" (where
N is a number). This is not a valid Java
identifier (causing an issue for PythonGenerator
which create field names from the variable names),
so the names are santized before going into Java.