Store both Java and PythonLikeObject version of Objects in PythonList
PythonList extends PythonLikeList so it can be used in translated
bytecode. Because it extends PythonLikeList, all of its contents
must be PythonLikeObjects. This isn't the case for list of ints,
String, or booleans (which are Java Integer, Boolean and String).

To fix, two seperate lists are maintained: the
cachedPythonLikeObjectList (which is the delegate used by
PythonLikeList), which only have PythonLikeObjects in it, and
cachedObjectList, which contains the Java Objects that are
expected by annotations.