Split types package into multiple subpackages; remove lambda wrappers for PythonLikeFunction
The types package was getting large, making it hard to find particular
classes. Numeric types (including PythonBoolean) now are in the numeric subpackage,
collection types are in the collections subpackage, and wrapper of
python and Java objects are in the wrappers package.

Lambda wrappers for PythonLikeFunction are no longer used thanks to
PythonOverloadImplementor creating a custom + optimized function
for dispatching to different signatures and handling different types.
This also allow the removal of PythonNumericOperations.