Add missing set methods and add frozenset to bytecode translator
To assist in testing, FunctionVerifier in conftest was created, and
an instance of it can be obtained by `verifier_for(function)`. It
ensures the Java implementation matches the Python implementation.
Additionally, if expected_result or expected_error is specified,
the Python implementation is checked to ensure it also return/throw
the specified value.