Support arbitary PythonComparable as @planning_id
Due to OptaPlanner sanity checking @PlanningId, the field/method
must return a Comparable object. Thus, for methods annotated
with @planning_id, their corresponding field is of type
Comparable. This does not pose an issue for int and str,
which have correspond Java objects. It does pose an issue
for custom Python objects, like uuid, since you cannot
assign an OpaquePythonReference to Comparable. To fix this,
when getting attributes during cloning, _get_python_object_attribute
will return a PythonComparable if the attribute getter was annotated
with `@planning_id`, and otherwise return an OpaquePythonReference.