Fix empty module in native check; do not try to convert a function to a class instance
Certain objects do not have a module (which in the function to get the
module, is indicated by ''). These objects are assumed to not be native,
to allow the translator to convert them.

Previously, if the type function is in globals, then the translator will
incorrectly try to convert user functions as an "instance" of the
native function class instead of translating its bytecode. Now,
before creating instances of a class, we first check if it is
a user function (and if so, do not create an instance, so that
it can be translated).