Add string operations and tests; force all types to be initialized before generating dispatches
A bit of trivia on Java static constants. Java static constants
are initialized when the their declaring class is initialized.
A class is initialized when one of the following occurs:

- T is a class and an instance of T is created.
- A static method declared by T is invoked.
- A static field declared by T is assigned.
- A static field declared by T is used and the field is not a constant variable
- T is a top level class and an assert statement lexically nested within T is executed.

(see https://docs.oracle.com/javase/specs/jls/se8/html/jls-12.html#jls-12.4.1)

This typically works okay. However, if you have classes A, B, which
refer to each other, and it is not the case (A instanceof B) or
(B instanceof A), then the order of initialization is undefined.

Before, this cause an annoying issue:

- The PythonLikeType for a particular class is defined in it
  corresponding class

- That PythonLikeType refer to PythonLikeTypes in DIFFERENT
  classes when declaring dispatches

- The PythonLikeType(s) in the DIFFERENT classes refer to
  the original PythonLikeType when declaring their dispatches.

This create an issue where one of the PythonLikeTypes will be
null, causing a NullPointerException.

To fix this, all PythonLikeType for builtin types are moved to
the class BuiltinTypes, so it is guarateeded they are not null
when dispatches are generated.