Support ValueRange, CountableValueRange and entity @value_range_provider
There was a bug with @value_range_provider where it would only work
without @problem_fact_collection_property/@planning_entity_collection_property
if it was for a OpaquePythonReference type due to its generated function signature.

Now the generated function signature is:
- Type, if ValueRange is assignable from Type
- List<Type>, if Type has a corresponding Java class
- OpaquePythonReference[] otherwise

A ValueRange cannot be directly assigned to the value range's field
in the PlanningEntity class because it is proxied by JPype. So we
create a NEW Proxy (one that implements ValueRange/CountableValueRange),
that reuses its InvocationHandler, which is used instead (and can
be cast to ValueRange as a result).