Add tests for try-except; fix raise Type and raise e from c impl
The implementor for raise e incorrectly assumed e must be an
exception, when it can be an exception or type. It will now
perform an instanceof check to determine if it need to create
an instance first.

The implementor for raise e from c did not add a DUP call before
the instanceof, causing the stack to be empty and causing
bytecode verification errors. The DUP was added, and
PythonBaseException will now set the __cause__ attribute
when the cause is set.

Add a method for getting the dict of a python reference
as a Map<String, PythonLikeObject>, which allow the
Java interpreter to set attributes of an exception
(and other classes that do not extend CPythonBackedObject)
when objects are shared between CPython and the Java intepreter.