Make translated class implement Comparable if it implement any comparison methods
Python does not have a generic compareTo method like Java. Instead,
it has individual methods for less than, greater than, less than or
equal, and greater than or equal. Strictly speaking, these methods
do not actually have a contract (that is, it is valid in Python
for a < b == 10, or to have a < b and a > b). However, it is
usually bad practice to not follow the normal comparison laws.

The code implement compareTo from one of these comparison
operations by implementing the following code:

    public int compareTo(Object other) {
        if (!(other instanceof this.class)) {
            throw new NotImplementedError();
        }

        if (this.lessThan(other)) {
            return -1;
        } else if (other.lessThan(this)) {
            return 1;
        } else {
            return 0;
        }
    }

To implement the above code with greaterThan, lessThanOrEqual, and
greaterThanOrEqual, the below properties for comparisons are used:

- a.greaterThan(b) == b.lessThan(a)
- !a.lessThanOrEqual(b) == a.greaterThan(b)
- !a.greaterThanOrEqual(b) == a.lessThan(b)