Add tests and fix bugs for methods in classes; refactor how builtin methods work
Move builtin methods from ObjectBuiltinOperations to PythonLikeObject to
allow them to be overriden. Previously, if a builtin method was
overriden, and the best known type is PythonLikeObject, then
the base method is always called (ignoring the override) since
it calls the static method in ObjectBuiltinOperations. Having
it in PythonLikeObject allow it to be overriden.

- Fix @classmethod methods so they are called with the type.
- Add a __getattribute__ override in type that always call the
  descriptor to match CPython behavior.
- Add support for returning a new CPythonBasedObject that
  was creating in Java (and thus does not have a corresponding
  CPython instance) to CPython.
- Set builtin type attributes (__name__ and co) when class is built
- Override __getType in dispatchers to return the corresponding function
  type (one of CLASS_FUNCTION_TYPE, STATIC_FUNCTION_TYPE and
  FUNCTION_TYPE), so the correct __get__ descriptor get called.