Add parent list to end of super MRO list to preserve local precedent order
In the C3 Linearization algorithm, the list of parents is included as
the last element of the parent's MRO list (despite not being a MRO
in itself), to keep precedent order as declared by a child. For
instance, for Child(A,B,C), we want B to appear after A, and
C to appear after B, in the linerization. Having the list of
parents last guarateeds that ordering, since it prevents a later
parent from being picked before all previous parents are picked.