Add tests for inner functions, fix bound variables that are free variables in inner functions
If a bound variable is a free variable in an inner function, it will not
appear in co_varnames. The previous code mapping bound variables to
their slots simply used the index of the bound variable in co_varnames,
but since it will be null for bound variables that are free in an
inner function, it will cause the generated code to be incorrect.

To fix it, if the bound variable is not found in co_varnames, it
is given the slot pythonCellVariablesStart + i (where i is the
index of the cell variable).