jpyinterpreter: wrap untranslated code objects so functions can use them
Translated functions expect a PythonCode object when MAKE_FUNCTION
is used. Previously, code objects that could not be translated
were wrapped in PythonObjectWrapper, which causes function that
use the code object with MAKE_FUNCTION to throw a ClassCastException.
To fix this, if a code object failed to be translated, we create
a Java function that wraps the code object and calls it. We then
wrap that Java function in a PythonCode object. When MAKE_FUNCTION
is called with a wrapped python code object, the Java function's
constructor calls CPython to create the corresponding Python
function for that code object.