Fix PythonInteger and PythonFloat __mod__ to copy CPython's behavior
In CPython, the sign of the remainder is always the same as the sign
of the right hand operand. To convert the Java remainder to the
Python remainder, if the Java remainder have a different sign
than the Python remainder, add the divisor to the remainder.
Additionally, now ZeroDivisionError is raised if a division by
0 is attempted. Also added missing tests for __mod__.