Support INPLACE operations delegating to their immutable counterpart
In CPython, if a class does not define an inplace operation, or
the inplace operation return NotImplemented, it will be delegated
to the immutable counterpart. For example, if __iadd__ is not
defined or returns NotImplemented, it will try __add__, and
if that not defined or return NotImplemented, it will try
__radd__ and fail with a TypeError if __radd__ is not
defined or return NotImplemented.