Fix native modules, add tests for modules, allow access to datetime fields
Because 'builtins' was not a banned module, many things that should not
be converted like the '<builtin function-or-method>' type were converted
to an empty class. This cause builtin functions and methods to become an
instance of the empty class, meaning calling them will always return
None. 'builtins' was added to the banned module list, which prevents
this issue.

Extracted is_native_module from is_c_native so it can be used outside
of is_c_native.

Detected untranslated Python Objects in LOAD_GLOBAL will no longer
throw an Error. In the future, FunctionMetadata will have a field
to store errors, warnings, and other things to note when
translating the Python function, and it will set a flag on it
(which then can be read after compilation is finished to decide
if they want to use the compiled function or not).