Created ArgumentSpec to easily create PythonLikeFunction from java function
Also started adding tests for date, and convert java datetime objects to
python datetime objects when transfering from java to python.

ArgumentSpec contains all the needed info to create the argument list
from positional and keyword arguments. The required info for the
argument spec is:

- Is the argument a positional varg, keyword varg, positional only,
  keyword only, or can be either positional/keyword?
- Does the argument have a default? If so, what the value?
- What the argument name?
- What the argument expected type?

ArgumentSpec is a fluent interface for entering all the info,
and provides methods that converts an N-arg function to
a PythonLikeFunction, with Keyword handling, argument type checking,
and checking for missing/extra arguments (and storing extra
arguments into the varg/kwargs if the function has one).