Modified how defaults, vargs and kwargs work in jpyinterpreter
Previously, defaults, vargs, kwargs worked in the following way:

- If the function signature is known:
    - CALL_METHOD: Cast parameters to method parameter types, and call
      the method

    - CALL_FUNCTION_KW: Create a defaults object with the keyword
      argument tuple, which load keyword arguments in reverse
      and positional arguments forward. The argument are stored
      in fields, and the argument have their default value set
      in the constructor.

    - CALL_FUNCTION: Same as CALL_FUNCTION_KW, but first loads
      an empty tuple to TOS to act as the empty keyword argument

    - UNPACK: handled as function signature unknown

- If the function signature is unknown:
    CALL_METHOD, CALL_FUNCTION_KW, CALL_FUNCTION and UNPACK
    calls the dispatcher, which finds a corresponding signature
    by argument count. If it cannot find a signature, it will
    use the GenericPythonFunctionSignature, which does not have
    any default supports and the function must take a List and
    a Map.

There were a few problem with this approach:

- The code for vargs and normal function calls are seperate
- vargs and kwargs functions can have defaults

There was also a bug with the Defaults class: it should
read positional arguments in reverse, since the last
positional argument is at the TOS.

The new way defaults, vargs and kwargs are handled:

- PythonFunctionSignature now have optional varg and kwarg
  variable index, which tells it where to put any extra
  positional and keyword arguments. This removes the need
  for GenericPythonFunctionSignature.

- Defaults is now responsible for populating vargs and kwargs in
  the known signature case.

- Defaults now get positionalArgumentIndex (= total args - number of
  keyword args - 1) in its constructor, which allow it to populate
  vargs.

Additionally, how PythonLikeFunction initializes its parameters
changed. Before, it used FunctionImplementor.extractArguments,
which is unwieldly to use. Now it use ArgumentSpec, which is
easier to use.

The Function generated class has two new fields:

- __spec_getter__ (static), which is a bifunction that takes the positional and
  keyword defaults and return the corresponding ArgumentSpec for the
  function

- __spec__ (instance), which is the ArgumentSpec for the function
  instance (with its defaults and keyword defaults, which may
  be different per instance).

PythonCompiledFunction has the following new fields:

- defaultPositionalArguments, which store (known) default positional
  arguments

- defaultKeywordArguments, which store (known) default keyword
  arguments

- co_posonlyargcount, the number of positional only arguments
  (that cannot be set using keywords)

- supportExtraPositionalArgs, true if vargs is supported, false
  otherwise

- supportExtraKeywordsArgs true if kwargs is supported, false otherwise
  (not new, just noting it)

- methodKind, store the kind of function it is (virtual, class or static)

There was a bug in the old code: it used co_argcount in many places.
However that is not the true argument count for the function;
it does not include keyword only arguments, nor the vargs
or kwargs argument. The field is now "encapsulated"
(still public so Python can access it), and code that used
it before now use totalArgCount(), a new function which
return the actual parameter count.

Simple recursion case (a function that calls itself) now
only generate a single function instead of many functions
(stoping when the stack overflows). The function is stored
in its cell variables, which when converted, calls the
compile python function, causing recursion in the compile
(since it will do it again when copying the closure). Now
when copying the closure, if it detects the function that
being compiled, it is replaced by a recursion marker, which
is replaced with the actual function instance when the
function instance is created.

Mutual recursion still generate many versions of the
function, but, it appears to work despite the stack
overflow.

Already created function instances for given interfaces
are now cached and reused in a dict. Defaults were not
working properly without this for some reason.