Put an empty function wrapper into instance map before translating function
Putting an empty wrapper into the instance map before translating the
function allows functions to reference each other without causing
a stack overflow (since it will see the instance map has the wrapper,
and use it, instead of trying to compile the already compile function).

This also allows us to remove the code for the self recursion, which
instead will use the empty wrapper trick.

The cached interface to function dicts cannot be removed; they
are needed for defaults.