Combine all ArgumentSpec into a single class; correctly handle class methods in dispatcher
PythonLikeType have "addMethod" methods for adding a
PythonFunctionSignature, which should always be used so a dispatcher
function can be generated (instead of setting an attribute
with __setAttribute). As such, the asPythonLikeFunction methods
of ArgumentSpec is an anti pattern, since it allows using the
bad method (using __setAttribute). Since ArgumentSpec will only
be used to convert to PythonFunctionSignature and to extract
arguments, the many variants of ArgumentSpec (ZeroArgumentSpec to
SixteenArgumentSpec) can be removed (and the extra function
interfaces created for them cna also be removed).

PythonLikeType's "getMethodKind" did not handle methods added
with "addMethod" correct, since it ignores function signatures
and only check attributes. It now check the overload list to
correctly determine the function type.

The TypeConvertor was missing a case for the Type Type, causing
it to return as the Function Type instead of the Type Type.

PythonFunctionSignature will now unwrap the bound method for
Class functions.