jpyinterpreter: Add tests for TimeDelta, generate constructor overload
PythonOverloadImplementor did not generate a overload for a known
constructor function before (since it uses a different API to control
the $call method for PythonLikeType).

divmod did not handle cases where a type does not define
a __divmod__ function. Fix it to use __floordiv__ and
__mod__ if __divmod__ and __rdivmod__ is not available.

Removed no longer required null checks for keywordArgs in
GlobalBuiltins.