Add tests for datetime; set type argument in argument holder for class methods
The code that generate method calls left the type argument on the stack
as if it was the instance for a virtual method. This is incorrect,
since it will be passed as an argument. Now for (class methods
only), it will set the corresponding field in defaults (for
virtual method, the above behavior is correct).