Add tests for set and dict comprehensions
The docs for SET_ADD and LIST_APPEND in dis are
false: their argument is distance from
TOS, not TOS1. MAP_ADD docs in dis are true.
As a result, MAP_ADD need to push the arg + 1
stack item to the TOS, whereas SET_ADD and
LIST_APPEND Need to push the arg stack item to
the TOS.