optapy#107: Add ConstraintVerifier API
Because solution cloning does not occur in the ConstraintVerifier
API (because OptaPlanner will not change the given solution/facts),
wrapping a fact will cause non-given facts that it reference to
have null fields (as wrapping does not go through the entire
convert_to_python_like_object process). Thus, after all
facts are wrapped, we need to iterate through the values of
the reference map (which will contain all facts, including
non-given facts) and update their fields from the Python
object.

Added test_solve_partial when I was debugging the cause; it
already passes with existing solve() code, so the
null field issue only affects the given(facts...) method of
ConstraintVerifier. Nonetheless, it is a useful test, so I kept it.