Add workflow to run SonarCloud on main
The majority of testing is done on in Python tests;
the Java tests mostly act as a sanity check, since
in order to actually test all the parts, we need to
test in Python (otherwise, we are just testing
how we think it works, not how it actually works).

In order to let Python tests contribute to the
test coverage, the action downloads the JaCoCo
runtime agent, and add it as an argument when
the test JVM starts.