Add Python test coverage
Like the tutorial in
https://docs.sonarcloud.io/enriching/test-coverage/python-test-coverage/,
we use Pytest with Coverage.py. Coverage.py seem to be confused
by --import-mode=importlib, which we use because JPype have an
import hook for Java classes. It causes the files and packages to
reference the tox virtual environment packages/classes, instead of the
source packages/classes.

For instance, For instance, for a package, it will report:

<package name=".tox.py39.lib.python3.9.site-packages.jpyinterpreter" ...>

and for a file it will report:

<class name="jvm_setup.py" filename=".tox/py39/lib/python3.9/site-packages/jpyinterpreter/jvm_setup.py" ...>

We need to convert the package and file names into their true names; i.e. for the example above:

<package name="jpyinterpreter" ...>
<class name="jvm_setup.py" filename="jpyinterpreter/src/main/python/jvm_setup.py" ...>

fix-coverage-paths.py is a Python script that does the above transformation for us and merge the two
separate coverage files into one.