Copy values view before iterating the value view in ConstraintVerifier.given
After fixing MirrorWithExtrasMap.values() method
so it conforms to the Map API (the Collection
MUST reflect changes to the Map and vice versa),
converting facts in ConstraintVerifier raise
a ConcurrentExecutionException since it iterating
through the changing map.

To prevent the ConcurrentExecutionException, we
copy values to it own list, and iterate through
that list instead.