Fix Python 3.11 calls; add instance field type info in PythonGenericType
PythonGenericType is a type of a particular type, and thus has access
to that type's static and class methods (which can be accessed without
an instance). Overriden getInstanceFieldDescriptor() in PythonGenericType
to return a fake FieldDescriptor that describe the method object which
will be on the type. It is fake in the sense the Java object does
NOT have a corresponding field for it, but it exists in the Python
attributes dictionary.

Add the missing Type -> Constructor conversion in the Python 3.11
CALL opcode and implementor (which allows constructors to be
specialized).

Add a NULL type, so it is easier to see where NULL are on the stack
when debugging.

PythonKnownFunctionType returns BASE_OBJECT for __getType and
__getGenericType, since accessing it via attribute returns an
object, not a type.

Python 3.11 calls did not use the correct indicies for defaults
(they used the argument index, not the default index (which is
less if not all arguments have default values). It also did
not pop off NULL on the stack. The code for poping off NULL
and the function is moved to PythonFunctionSignature instead of
FunctionImplementor, since we need to know the function kind to
determine if we need to pop off the function off the stack.