optapy#143: Fix truncating floats to 32-bits in conversion (#144)
JFloat was used instead of JDouble when converting a CPython
float (which is 64-bits) to a jpyinterpreter float, causing
the value to be incorrectly truncated.

Fixing this issue exposed another issue: BigDecimal.valueOf(double)
returns the shortest BigDecimal that has the same double value,
not the canonical double value. This means for `1.055`, which
is actually stored as `1.0549999....`, when rounded to 2 decimal
places, returns `1.06` (incorrect) instead of `1.05` (correct).
This is fixed by using `new BigDecimal(double)` instead, which uses
the canonical double value.