optapy#150: Fix nullable variables and deleted attributes
The old generated wrapper for the setters did not consider that
valueAsPythonLikeObject can be null, and thus can potentially
pass null to the delegated setter. For simple getters/setters
that don't modify the parameter, it "deletes" the attribute
(since jpyinterpreter use null to mean "missing attribute").
Thus, we need to change the null to None if None is assignable
to the setter's parameter (in the case None is not assignable
to the parameter, we still pass null, since planning variable
getters/setters should be simple).

This was worsen by the fact `jpyinterpreter` assumes all known
attributes of a type will always exist, and thus cause null
to propagate through code instead of raising an AttributeError.
`jpyinterpreter` now checks if the field is null, and if so, throw
an AttributeError.

The test for custom shadow variables had typing errors (in particular,
the shadow variable and the planning variable are nullable int, but
was typed as int, meaning a TypeError occurs when you try to assign
None to them. I changed their types to Optional[int], which allows None
to be assigned to them.