optapy#155: Check if field is assignable to None before passing None to setter
The previous implementation only checked if the setter can accept None
as a parameter. However, it might be the case the user typed the
fields and not the setters, causing an ClassCastException. Now None
is passed to the setter only if both the field and setter allow it;
null (i.e. delete attribute) is passed otherwise.