optapy#154: Update all attributes of all wrapped objects after clone
Additionally, fix the type of problem facts and planning entities
that are subclassed.

Previously, some attributes escaped being updated when they are
not accessible from the planning solution class (ex: inside
the value range of a planning entity). To account for this,
after the entire solution is cloned, we go through the list
of wrapped objects and update them from their corresponding
object. This need to be done after everything is cloned, so
the reference map contains every discovered planning entity
and problem fact (if done in the planning entity setFields,
multiple instances of the same object may be constructed,
potentially causing an infinite loop in vehicle-routing).