optapy#163: Fix stub file paths so they are added to the wheel
When jpyinterpreter was merged, the project structure changed.
In particular, the optapy-core code moved from <project-root>
to <project-root>/optapy-core. Some of the stub files paths
in setup.py was not changed to reflect this, resulting in
the stubs files being missing in the distribution.
This fixes the paths so they are be relative to
<project-root>/optapy-core.

Additionally, we use buildpy from setuptools and remove the
distutils try...except, since distutils will be removed in
Python 3.12 (https://peps.python.org/pep-0632/).